Move calculateCenterFromCameraLngLatAlt to helper, remove redundant methods.
